internalize stream fd
