+++ src/ray/util/tests/stream_redirection_utils_test.cc
@@ -41,7 +41,6 @@ TEST(LoggingUtilTest, RedirectStderr) {
 
   StreamRedirectionOption opts;
   opts.file_path = test_file_path;
-  // Manually check whether streamed content is displayed on screen.
   opts.tee_to_stderr = true;
   opts.rotation_max_size = 5;
   opts.rotation_max_file_count = 2;
